# product-service/doc/api.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger Petstore"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "7537ljc0qf.execute-api.us-east-1.amazonaws.com"
basePath: "/dev"
tags:

- name: "store"
  description: "Access to our book shop"

schemes:
- "https"
- "http"
paths:
  /{products}:
    get:
      tags:
      - "books"
      summary: "Get all books"
      description: ""
      operationId: "getAllBook"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "products"
        in: "path"
        description: "Get all books"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Book"
        "400":
          description: "Invalid "
        "404":
          description: " not found"
  /products/{productById}:
    get:
      tags:
      - "books"
      summary: "Get user by user name"
      description: ""
      operationId: "productById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "productById"
        in: "path"
        description: "Get one book "
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Book"
        "400":
          description: "Invalid "
        "404":
          description: " not found"


definitions:
  Book:
    type: "object"
    properties:
      id:
        type: "string"
  
      title:
        type: "string"
      year:
        type: "string"
      author:
        type: "string"
      count:
        type: "number"
      description:
        type: "string"
      price:
        type: "number"
      picture:
        type: "string"
      
    xml:
      name: "User"



externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"